Organize component files in folders
